Drop standalone override CLI and unused UC volume I/O.
Keep override helpers for the onboard job only: local-path batch apply
via gate_2. Restore Streamlit unity_volume_files to the develop version.
