+++ src/edvise/genai/mapping/schema_mapping_agent/manifest/hitl/__init__.py
@@ -15,14 +15,9 @@ from edvise.genai.mapping.schema_mapping_agent.manifest.hitl.artifacts import (
 from edvise.genai.mapping.schema_mapping_agent.manifest.hitl.override import (
     ManifestMappingOverrideRequest,
     ManifestOverrideError,
-    load_correction_json,
     load_overrides_json,
     override_manifest_mapping,
-    override_manifest_mapping_at_path,
-    override_manifest_mapping_on_volume,
     override_manifest_mappings,
-    override_manifest_mappings_at_path,
-    override_manifest_mappings_on_volume,
     unmapped_field_mapping_record,
 )
 from edvise.genai.mapping.schema_mapping_agent.manifest.hitl.resolver import (
@@ -47,14 +42,9 @@ __all__ = [
     "HITL_CONFIDENCE_THRESHOLD",
     "ManifestMappingOverrideRequest",
     "ManifestOverrideError",
-    "load_correction_json",
     "load_overrides_json",
     "override_manifest_mapping",
-    "override_manifest_mapping_at_path",
-    "override_manifest_mapping_on_volume",
     "override_manifest_mappings",
-    "override_manifest_mappings_at_path",
-    "override_manifest_mappings_on_volume",
     "unmapped_field_mapping_record",
     "InstitutionSMAHITLItems",
     "SMAFailureMode",
