+++ src/edvise/genai/mapping/streamlit-genai-hitl-app/hitl_reviewer/platform/unity_volume_files.py
@@ -1,24 +1,56 @@
 """
-Read and write file contents on Unity Catalog volumes via the Databricks Files API.
+Read and write file contents on Unity Catalog volumes via the Databricks Files API
+(``/api/2.0/fs/files/...``). Mirrors :class:`databricks.sdk.service.files.FilesAPI`.
 
-Re-exports :mod:`edvise.genai.mapping.shared.unity_volume_files` so the Streamlit app
-can keep its existing import path. Typical GenAI HITL paths are on the institution
-**silver** volume, under ``.../genai_mapping/runs/...`` (same layout as
-``edvise_genai_ia`` / ``edvise_genai_sma``).
+Typical GenAI HITL paths are on the institution **silver** volume, under
+``.../genai_mapping/runs/...`` (same layout as ``edvise_genai_ia`` / ``edvise_genai_sma``).
 """
 
 from __future__ import annotations
 
-from edvise.genai.mapping.shared.unity_volume_files import (
-    is_unity_catalog_volume_path,
-    read_unity_file_bytes,
-    read_unity_file_text,
-    write_unity_file_text,
-)
+import io
+from typing import cast
 
-__all__ = [
-    "is_unity_catalog_volume_path",
-    "read_unity_file_bytes",
-    "read_unity_file_text",
-    "write_unity_file_text",
-]
+from databricks.sdk import WorkspaceClient
+from databricks.sdk.core import Config
+
+
+def _client() -> WorkspaceClient:
+    return WorkspaceClient(config=Config())
+
+
+def read_unity_file_bytes(absolute_path: str) -> bytes:
+    """
+    Read a file at a UC volume (or other Files-API) absolute path, e.g.
+    ``/Volumes/catalog/schema/vol/path/identity_grain_hitl.json``.
+    """
+    path = str(absolute_path).strip()
+    if not path or not path.startswith("/Volumes/"):
+        raise ValueError(
+            "Path must be a non-empty Unity Catalog /Volumes/… absolute path"
+        )
+    c = _client()
+    out = c.files.download(path)
+    if out.contents is None:
+        return b""
+    return cast(bytes, out.contents.read())
+
+
+def read_unity_file_text(absolute_path: str, *, encoding: str = "utf-8") -> str:
+    return read_unity_file_bytes(absolute_path).decode(encoding)
+
+
+def write_unity_file_text(
+    absolute_path: str, text: str, *, encoding: str = "utf-8", overwrite: bool = True
+) -> None:
+    path = str(absolute_path).strip()
+    if not path or not path.startswith("/Volumes/"):
+        raise ValueError(
+            "Path must be a non-empty Unity Catalog /Volumes/… absolute path"
+        )
+    c = _client()
+    c.files.upload(
+        path,
+        io.BytesIO(text.encode(encoding)),
+        overwrite=overwrite,
+    )
